updated files for PR comments
