+++ app/processing/delta/index.js
@@ -1,10 +1,10 @@
 const allocateToLedgers = require('./allocate-to-ledgers')
 const calculateLineDeltas = require('./calculate-line-deltas')
 const calculateOverallDelta = require('./calculate-overall-delta')
+const getDefaultLedger = require('./get-default-ledger')
 const getInvoiceLines = require('./get-invoice-lines')
 const getOutstandingLedgerValues = require('./get-outstanding-ledger-values')
 const zeroValueSplit = require('./zero-value-split')
-const copyPaymentRequest = require('./copy-payment-request')
 
 const calculateDelta = (paymentRequest, previousPaymentRequests) => {
   const invoiceLines = getInvoiceLines(paymentRequest, previousPaymentRequests)
@@ -30,6 +30,16 @@ const calculateDelta = (paymentRequest, previousPaymentRequests) => {
   }
 
   return { deltaPaymentRequest, completedPaymentRequests: [updatedPaymentRequest] }
+}
+
+const copyPaymentRequest = (paymentRequest, overallDelta, lineDeltas) => {
+  return {
+    ...paymentRequest,
+    value: overallDelta,
+    settledValue: 0,
+    ledger: getDefaultLedger(overallDelta),
+    invoiceLines: lineDeltas.filter(x => x.value !== 0)
+  }
 }
 
 module.exports = calculateDelta
