+++ app/processing/route-manual-ledger-to-request-editor.js
@@ -6,12 +6,13 @@ const util = require('util')
 
 const routeManualLedgerToRequestEditor = async (paymentRequest) => {
   const transaction = await db.sequelize.transaction()
+  const { deltaPaymentRequest, completedPaymentRequests } = paymentRequest
   try {
-    const manualLedgerMessage = { paymentRequest: paymentRequest.deltaPaymentRequest, paymentRequests: paymentRequest.completedPaymentRequests }
+    const manualLedgerMessage = { paymentRequest: deltaPaymentRequest, paymentRequests: completedPaymentRequests }
     await sendManualLedgerMessage(manualLedgerMessage)
     console.log('Payment request routed to request editor for manual ledger check:', util.inspect(manualLedgerMessage, false, null, true))
-    const holdCategoryId = await getHoldCategoryId(paymentRequest.deltaPaymentRequest.schemeId, 'Manual ledger hold', transaction)
-    await holdAndReschedule(paymentRequest.deltaPaymentRequest.schemeId, paymentRequest.deltaPaymentRequest.paymentRequestId, holdCategoryId, paymentRequest.deltaPaymentRequest.frn, transaction)
+    const holdCategoryId = await getHoldCategoryId(deltaPaymentRequest.schemeId, 'Manual ledger hold', transaction)
+    await holdAndReschedule(deltaPaymentRequest.schemeId, deltaPaymentRequest.paymentRequestId, holdCategoryId, deltaPaymentRequest.frn, transaction)
     await transaction.commit()
   } catch (error) {
     await transaction.rollback()
